+++ views/config/userGoogleAuthenticator.php
@@ -8,15 +8,15 @@
 
 use humhub\modules\twofa\drivers\GoogleAuthenticatorDriver;
 use humhub\modules\twofa\models\GoogleAuthenticatorUserSettings;
-use humhub\widgets\Button;
-use yii\bootstrap\ActiveForm;
+use humhub\widgets\bootstrap\Button;
+use humhub\widgets\form\ActiveForm;
 
 /* @var $driver GoogleAuthenticatorDriver */
 /* @var $form ActiveForm */
 /* @var $model GoogleAuthenticatorUserSettings */
 /* @var $requirePinCode bool */
 ?>
-<div id="twofaGoogleAuthCode" class="form-group">
+<div id="twofaGoogleAuthCode" class="mb-3">
     <?= $driver->getQrCodeSecretKeyFile(['requirePinCode' => $requirePinCode]) ?>
 </div>
 
@@ -25,7 +25,7 @@ use yii\bootstrap\ActiveForm;
     <?= $form->field($model, 'changeSecretCode')->hiddenInput()->label(false) ?>
 </div>
 
-<div class="form-group">
+<div class="mb-3">
     <?= Button::asLink(Yii::t('TwofaModule.base', 'Request new code'))
         ->icon('fa-qrcode')
         ->right()