+++ docs/CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog
 - Enh #66: Reduce translation message categories
 - Enh #81: Time Based Code
 - Enh #78: Migration to Bootstrap 5 for HumHub 1.18
+- Enh #71: Time Based Code
+- Fix #76: Fix timezone in email
 
 1.1.3 (June 8, 2023)
 --------------------