+++ docs/CHANGELOG.md
@@ -1,11 +1,13 @@
 Changelog
 =========
 
-1.1.4 (Unreleased)
-----------------------
+1.2.0 (unreleased)
+-------------------------
 - Enh #58: Tests for `next` version
 - Enh #65: Use PHP CS Fixer
 - Enh #66: Reduce translation message categories
+- Enh #81: Time Based Code
+- Enh #78: Migration to Bootstrap 5 for HumHub 1.18
 - Enh #71: Time Based Code
 - Fix #76: Fix timezone in email
 - Fix #80: Update test user passwords